Pass $event to rcp_stripe_charge_succeeded hook
I don't _think_ there's any downside to this, it should be fully backwards compatible, and it allows for using the stripe event data to inform what actions an extension might want to take. Probably a bit of an edge case in terms of usefulness, but it'd be handy for something I'm working on right now.